requirements.txt
torch
torchvision
transformers
pycocotools
pillow
pyyaml
tqdm